backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-sdk-python>=0.1.0",
    "fastapi>=0.115.14",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=1.10.22",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
